More tweaks to the photowall script.

+++ make-photo-wall.py
@@ -1,52 +1,48 @@
 #!/usr/bin/env python
+
+import hashlib
 import os
-import sqlalchemy
-import oculum
-import shelve
-import time
+import random
+import urllib
+
 import sh
 import bs4
-import urllib
-import hashlib
 import requests
 
-# For development purposes
-datadir = oculum.datadir
-gravatar_dir = oculum.gravatar_dir
-montage_dir = oculum.montage_dir
+dimensions = (12, 5)
 
-try:
-    os.makedirs(datadir)
-except OSError:
-    pass
-
-try:
-    os.makedirs(gravatar_dir)
-except OSError:
-    pass
-
-try:
-    os.makedirs(montage_dir)
-except OSError:
-    pass
+datadir = './data'
+avatar_dir = datadir + '/avatars'
+montage_dir = datadir + '/montage'
 
 
-# We need 12*5==60 images to do this right.
-def avatars(N=12 * 5):
+def make_directories():
+    try:
+        os.makedirs(datadir)
+    except OSError:
+        pass
 
-    count = 0
+    try:
+        os.makedirs(avatar_dir)
+    except OSError:
+        pass
 
+    try:
+        os.makedirs(montage_dir)
+    except OSError:
+        pass
+
+
+def avatars(N):
     url = 'https://badges.fedoraproject.org/badge/mugshot/full'
-
     response = requests.get(url)
     soup = bs4.BeautifulSoup(response.text)
     last_pane = soup.findAll(attrs={'class': 'grid-100'})[-1]
     persons = last_pane.findAll('a')
-    for person in persons:
-        count = count + 1
-        if count >= N:
-            break
 
+    persons = random.sample(persons, N)
+
+    for person in persons:
         name = person.text.strip()
         openid = 'http://%s.id.fedoraproject.org/' % name
         hash = hashlib.sha256(openid).hexdigest()
@@ -55,22 +51,26 @@ def avatars(N=12 * 5):
 
 
 def make_montage(candidates):
-    """ Pull down gravatars to disk and stich with imagemagick """
+    """ Pull down avatars to disk and stich with imagemagick """
 
     filenames = []
     for name, url in candidates:
-        print "Grabbing", name, "at", url
-        filename = os.path.join(gravatar_dir, name)
-        urllib.urlretrieve(url, filename=filename)
+        filename = os.path.join(avatar_dir, name)
+        if not os.path.exists(filename):
+            print "Grabbing", name, "at", url
+            urllib.urlretrieve(url, filename=filename)
+        else:
+            print "Already have", name, "at", filename
         filenames.append(filename)
 
     args = filenames + [montage_dir + '/montage.png']
-    sh.montage('-tile', '12x5', '-geometry', '+0+0', *args)
+    sh.montage('-tile', '%ix%i' % dimensions, '-geometry', '+0+0', *args)
 
 
 def main():
-    candidates = avatars()
-
+    make_directories()
+    N = dimensions[0] * dimensions[1]
+    candidates = avatars(N)
     make_montage(candidates)
 
 if __name__ == '__main__':
